+++ src/app/beer_garden/db/mongo/util.py
@@ -229,6 +229,17 @@ def check_indexes(document_class):
             )
             raise
 
+    try:
+        if document_class.objects.count() > 0:
+            document_class.objects.first()
+        logger.warning("%s table looks good", document_class.__name__)
+    except (FieldDoesNotExist, InvalidDocumentError):
+        logger.error(
+            "%s table failed to load properly to validate old indexes and fields, please check the Change Log for any major model changes",
+            document_class.__name__,
+        )
+        raise
+
 
 def _update_request_parent_field_type():
     """Change GenericReferenceField to ReferenceField"""