+++ src/app/test/db/mongo/util_test.py
@@ -2,6 +2,7 @@
 import pytest
 from mock import MagicMock, Mock, patch
 from mongoengine import connect
+from mongoengine.errors import FieldDoesNotExist
 
 import beer_garden.db.mongo.models
 import beer_garden.db.mongo.util
@@ -15,9 +16,15 @@ from beer_garden.errors import IndexOperationError
 
 @pytest.fixture
 def model_mocks(monkeypatch):
-    request_mock = Mock()
-    system_mock = Mock()
-    job_mock = Mock()
+    request_mock = Mock(
+        objects=Mock(count=Mock(return_value=1), first=Mock(return_value=[{}]))
+    )
+    system_mock = Mock(
+        objects=Mock(count=Mock(return_value=1), first=Mock(return_value=[{}]))
+    )
+    job_mock = Mock(
+        objects=Mock(count=Mock(return_value=1), first=Mock(return_value=[{}]))
+    )
 
     request_mock.__name__ = "Request"
     system_mock.__name__ = "System"
@@ -132,6 +139,23 @@ class TestCheckIndexes(object):
 
         for doc in model_mocks.values():
             with pytest.raises(IndexOperationError):
+                beer_garden.db.mongo.util.check_indexes(doc)
+
+    @patch("mongoengine.connect", Mock())
+    @patch("mongoengine.connection.get_db", MagicMock())
+    def test_unsuccessful_read_objects(self, model_mocks):
+        for model_mock in model_mocks.values():
+            model_mock.list_indexes = Mock(return_value=["index1"])
+            model_mock._get_collection = Mock(
+                return_value=MagicMock(
+                    index_information=Mock(return_value={"index1": {}})
+                )
+            )
+
+            model_mock.objects.first.side_effect = FieldDoesNotExist("")
+
+        for doc in model_mocks.values():
+            with pytest.raises(FieldDoesNotExist):
                 beer_garden.db.mongo.util.check_indexes(doc)
 
     @patch("mongoengine.connection.get_db")